letsplay_gpu: Make EGL context binding explicit
Before, the context would simply be automatically made current on the creating thread. The previous is "idiomatic" per se, but, in theory, an EGL context, with the proper safety guards (locking, ensuring only one thread at a time holds the context), can be shared across threads.

The current design of another project that ends up using the letsplay_gpu crate is currently setup like so (of notice, it basically mirrors my ideas for letsplay_runner_core..):

- main "thread" & tokio workers
- execution thread (is where EGL is held currently)
- reactor thread (reacts to events posted via a mpsc channel.. thinking about it, this could use the stdlib, which would probably even be faster since it's not all being wrapped via futures::block_on or whatever)
- encoding thread

Currently the execution thread does one GPU readback, which the reactor thread then does *again* to copy to a libav frame. Not very ideal at all, especially not when already rendering to a texture that CUDA can interop to without even copying to system RAM.