retro_frontend: Provide a dummy audio sample callback
Badly behaved cores like LRPS2 (which I needed to build myself thanks to retroarch maintainers having the effective IQ of a jelly donut) assume this callback will never be null.

Currently it's a stub; later on we *could* buffer internally and call the sample callback the frontend interface has. For now, I just wanted to get LRPS2 to boot.