retro_frontend: Implement analog support properly
Input devices are now given the index (used for sticks currently, but there is a "analog button" API as well). Our Analog RetroPad actually half follows this kind of wrong (we let the user press with a pressure and we report that pressure, the real implementation is a digital button until you request with the index. Oh well, it can be fixed, and I don't think it does any real harm anyways)