retro_frontend: Clean up software conversion
It was a bit messy before.

RGB565 support now no longer has a dependency on the "rgb565" crate. The only real con is that now it's a bit bit-hackery. If a high level way of doing it were to be considered I'd prefer a different crate that can do all three pixel formats (or we can abstract the bit hackery ourselves, since we really only need conversion from N formats -> XRGB8888.)

Also add support for RGB1555. I don't think this path will be taken often if at all, but it's there now.